Covert CodableCloudPropertyValue & CodableUserDefaultsPropertyListValue to use Codable
This way, users don't have to conform to a new protocol in order to
store it - anything conforming to Codable is supported.

If the type conforms to CloudPropertyListValue or
UserDefaultsPropertyListValue - the existing storage implementation is
used. Only types that **don't** conform to *PropertyListValue but **do**
conform to Codable use the Codable storage implementation (JSON
encoding/decoding)